+++ src/components/CloudHeader/__tests__/CloudHeaderList-test.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import CloudHeaderList from './CloudHeaderList';
-import CloudHeaderListItem from './CloudHeaderListItem';
+import CloudHeaderList from '../CloudHeaderList';
+import CloudHeaderListItem from '../CloudHeaderListItem';
 import { shallow, mount } from 'enzyme';
 
 describe('CloudHeaderList', () => {
